// src/lib/types.ts
export interface GeneratedInfo {
  lastName: string;
  firstName: string;
  birthday: string;
  phone: string;
  password: string;
  email: string;
}

export interface Country {
  code: string;
  name: string;
  phonePrefix: string;
}

export interface EmailDomain {
  value: string;
  label: string;
}